backend/requirements.txt
boto3==1.35.10
fastapi==0.115.4
passlib==1.7.4
pydantic==2.9.2
PyJWT==2.3.0
python-dotenv==1.0.1
SQLAlchemy==2.0.31
uvicorn==0.32.0
psycopg2_binary==2.9.9
python-multipart
pydantic[email]